+++ packages/deriv_auth/lib/features/social_auth/cubit/social_auth_state.dart
@@ -1,7 +1,11 @@
 import 'package:deriv_auth/features/social_auth/models/social_auth_provider_model.dart';
+import 'package:equatable/equatable.dart';
 
 /// Abstract class to define all social authentication-related state.
-abstract class SocialAuthState {}
+abstract class SocialAuthState extends Equatable {
+  @override
+  List<Object?> get props => <Object?>[];
+}
 
 /// Initial state of [SocialAuthCubit].
 class SocialAuthInitialState extends SocialAuthState {}
@@ -18,6 +22,9 @@ class SocialAuthErrorState extends SocialAuthState {
 
   /// Error message.
   final String? message;
+
+  @override
+  List<Object?> get props => <Object?>[message];
 }
 
 /// Loaded state of [SocialAuthCubit].
@@ -29,4 +36,7 @@ class SocialAuthLoadedState extends SocialAuthState {
 
   /// List of social auth providers.
   final List<SocialAuthProviderModel> socialAuthProviders;
+
+  @override
+  List<Object?> get props => <Object?>[socialAuthProviders];
 }