+++ packages/deriv_widgetbook/android/app/src/main/kotlin/com/example/deriv_widgetbook/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.deriv.widgetbook
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}